example-02-duckhunt/machine-learning/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
torch
torchvision
Pillow
ultralytics
pandas
tqdm
pyyaml
requests
